backend/requirements.txt
fastapi
uvicorn
tensorflow
numpy
pydantic
python-multipart
httpx
joblib
python-dotenv
scikit-learn
